@@ -31,25 +31,51 @@
 		"svelte": "^5.0.0"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^6.0.0",
-		"@sveltejs/kit": "^2.16.0",
-		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.0",
-		"@tailwindcss/postcss": "^4.1.8",
-		"publint": "^0.3.2",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
-		"vite": "^6.2.6"
+		"@lucide/svelte": "^0.482.0",
+		"@sveltejs/adapter-auto": "^6.0.1",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/kit": "^2.21.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/typography": "^0.5.16",
+		"@tailwindcss/vite": "^4.1.7",
+		"@tanstack/table-core": "^8.20.5",
+		"@types/bun": "^1.2.14",
+		"@types/fs-extra": "^11.0.4",
+		"autoprefixer": "^10.4.21",
+		"layerchart": "2.0.0-next.6",
+		"svelte": "^5.33.0",
+		"svelte-check": "^4.2.1",
+		"sveltekit-superforms": "^2.19.1",
+		"tailwindcss": "^4.1.8",
+		"tslib": "^2.8.1",
+		"tw-animate-css": "^1.3.0",
+		"typescript": "^5.8.3",
+		"vite": "^6.3.5"
 	},
 	"keywords": [
 		"svelte"
 	],
 	"dependencies": {
-		"@sveltejs/adapter-static": "^3.0.8",
-		"autoprefixer": "^10.4.21",
-		"bits-ui": "^2.5.0",
+		"@types/js-search": "^1.4.4",
+		"bits-ui": "^1.5.3",
+		"clsx": "^2.1.1",
+		"embla-carousel-svelte": "^8.6.0",
+		"formsnap": "^2.0.1",
+		"fs-extra": "^11.3.0",
+		"js-search": "^2.0.1",
+		"lucide-svelte": "^0.511.0",
+		"mode-watcher": "0.5.1",
+		"noise_validator": "^1.2.21",
+		"paneforge": "^1.0.0-next.5",
+		"query-string": "^9.2.0",
+		"remove-markdown": "^0.6.2",
+		"sharp": "^0.34.2",
+		"svelte-legos": "^0.2.5",
+		"svelte-markdown": "^0.4.1",
+		"svelte-persisted-store": "^0.12.0",
+		"svelte-sonner": "^1.0.2",
 		"tailwind-merge": "^3.3.0",
-		"tailwindcss": "^4.1.8"
+		"tailwind-variants": "^0.2.1",
+		"vaul-svelte": "^1.0.0-next.7"
 	}
 }
