@@ -1,8 +1,8 @@
 import { LinkPreview as HoverCardPrimitive } from "bits-ui";
-
 import Content from "./hover-card-content.svelte";
+import Trigger from "./hover-card-trigger.svelte";
+
 const Root = HoverCardPrimitive.Root;
-const Trigger = HoverCardPrimitive.Trigger;
 
 export {
 	Root,
